Polish the interface for the date store factory
