--- src/providers/DateStore/makeDateStore.tsx
@@ -1,227 +1,0 @@
-import { createContext, type PropsWithChildren, useContext, useMemo, useState } from 'react'
-import {
-  endOfDay,
-  endOfMonth,
-  endOfYear,
-  max,
-  min,
-  startOfMonth,
-  startOfYear,
-} from 'date-fns'
-import { createStore, useStore } from 'zustand'
-
-import { unsafeAssertUnreachable } from '@utils/switch/assertUnreachable'
-import { useStoreWithDateSelected } from '@utils/zustand/useStoreWithDateSelected'
-
-export type DateSelectionMode = 'full' | 'month' | 'year'
-
-export function clampToAfterActivationDate(date: Date | number, activationDate: Date) {
-  return max([date, activationDate])
-}
-
-export function clampToPresentOrPast(date: Date | number, cutoff = endOfDay(new Date())) {
-  return min([date, cutoff])
-}
-
-export type DateRange = { startDate: Date, endDate: Date }
-type GetDateRangeOptions =
-  | { mode: 'full', startDate: Date, endDate: Date }
-  | { mode: Exclude<DateSelectionMode, 'full'>, startDate?: Date, endDate: Date }
-
-function getDateRange(options: GetDateRangeOptions): DateRange {
-  const mode = options.mode
-  switch (mode) {
-    case 'month':
-      return {
-        startDate: startOfMonth(options.endDate),
-        endDate: clampToPresentOrPast(endOfMonth(options.endDate)),
-      }
-    case 'year':
-      return {
-        startDate: startOfYear(options.endDate),
-        endDate: clampToPresentOrPast(endOfYear(options.endDate)),
-      }
-    case 'full':
-      return {
-        startDate: options.startDate,
-        endDate: clampToPresentOrPast(endOfDay(options.endDate)),
-      }
-    default:
-      unsafeAssertUnreachable({
-        value: mode,
-        message: 'Invalid mode',
-      })
-  }
-}
-
-function withCorrectedRange<TDateRange extends DateRange, TOut>(fn: (options: TDateRange) => TOut) {
-  return (options: TDateRange) => {
-    const { startDate, endDate } = options
-
-    if (startDate > endDate) {
-      return fn({ ...options, startDate: endDate, endDate: startDate })
-    }
-
-    return fn({ ...options, startDate, endDate })
-  }
-}
-
-export type DateState = DateRange
-
-type DateActions = {
-  setDate: (options: { date: Date }) => DateRange
-  setDateRange: (options: { startDate: Date, endDate: Date }) => DateRange
-  setMonth: (options: { startDate: Date }) => DateRange
-  setYear: (options: { startDate: Date }) => DateRange
-
-  setMonthByPeriod: (options: { monthNumber: number, yearNumber: number }) => DateRange
-}
-
-type DateStore = DateState & { actions: DateActions }
-
-function buildDefaultInitialRange(now: Date): DateRange {
-  return {
-    startDate: startOfMonth(now),
-    endDate: clampToPresentOrPast(endOfMonth(now)),
-  }
-}
-
-type MakeDateStoreOptions = {
-  buildInitialRange?: (now: Date) => DateRange
-}
-
-const getEffectiveDateForMode = (mode: DateSelectionMode, { date }: { date: Date }): { date: Date } => {
-  return { date: getDateRange({ mode, startDate: date, endDate: date }).endDate }
-}
-
-const getEffectiveDateRangeForMode = (
-  mode: DateSelectionMode,
-  { startDate, endDate }: { startDate: Date, endDate: Date },
-): { startDate: Date, endDate: Date } => {
-  return getDateRange({ mode, startDate, endDate })
-}
-
-export function makeDateStore({ buildInitialRange = buildDefaultInitialRange }: MakeDateStoreOptions = {}) {
-  function buildStore() {
-    const now = new Date()
-
-    return createStore<DateStore>((set) => {
-      const apply = (next: DateRange): DateRange => {
-        set(next)
-        return next
-      }
-
-      const setDate = ({ date }: { date: Date }): DateRange => {
-        // Always clamp to start of month for date.
-        const monthRange = getDateRange({ mode: 'month', endDate: date })
-        const fullRange = getDateRange({ mode: 'full', startDate: date, endDate: date })
-        return apply({ startDate: monthRange.startDate, endDate: fullRange.endDate })
-      }
-
-      const setDateRange = withCorrectedRange(({ startDate, endDate }): DateRange => {
-        return apply(getDateRange({ mode: 'full', startDate, endDate }))
-      })
-
-      const setMonth = ({ startDate }: { startDate: Date }): DateRange => {
-        return apply(getDateRange({ mode: 'month', endDate: startDate }))
-      }
-
-      const setYear = ({ startDate }: { startDate: Date }): DateRange => {
-        return apply(getDateRange({ mode: 'year', endDate: startDate }))
-      }
-
-      return {
-        ...buildInitialRange(now),
-
-        actions: {
-          setDate,
-          setDateRange,
-          setMonth,
-          setYear,
-
-          setMonthByPeriod: ({ monthNumber, yearNumber }) => {
-            const effectiveMonthNumber = Math.min(Math.max(monthNumber, 1), 12)
-
-            return setMonth({ startDate: new Date(yearNumber, effectiveMonthNumber - 1, 1) })
-          },
-        },
-      }
-    })
-  }
-
-  const DateStoreContext = createContext(buildStore())
-
-  function useDate({ dateSelectionMode = 'full' }: { dateSelectionMode?: DateSelectionMode } = {}) {
-    const store = useContext(DateStoreContext)
-
-    const rawDate = useStoreWithDateSelected(store, ({ endDate }) => endDate)
-
-    return useMemo(
-      () => getEffectiveDateForMode(dateSelectionMode, { date: rawDate }),
-      [dateSelectionMode, rawDate],
-    )
-  }
-
-  function useDateActions() {
-    const store = useContext(DateStoreContext)
-
-    const setDate = useStore(store, ({ actions: { setDate } }) => setDate)
-
-    return { setDate }
-  }
-
-  function useDateRange({ dateSelectionMode }: { dateSelectionMode: DateSelectionMode }) {
-    const store = useContext(DateStoreContext)
-
-    const rawStartDate = useStoreWithDateSelected(store, ({ startDate }) => startDate)
-    const rawEndDate = useStoreWithDateSelected(store, ({ endDate }) => endDate)
-
-    return useMemo(
-      () => getEffectiveDateRangeForMode(
-        dateSelectionMode,
-        { startDate: rawStartDate, endDate: rawEndDate },
-      ), [dateSelectionMode, rawEndDate, rawStartDate],
-    )
-  }
-
-  function useDateRangeActions() {
-    const store = useContext(DateStoreContext)
-
-    const setDateRange = useStore(store, ({ actions: { setDateRange } }) => setDateRange)
-    const setMonth = useStore(store, ({ actions: { setMonth } }) => setMonth)
-    const setYear = useStore(store, ({ actions: { setYear } }) => setYear)
-
-    return {
-      setDateRange,
-      setMonth,
-      setYear,
-    }
-  }
-
-  function usePeriodAlignedActions() {
-    const store = useContext(DateStoreContext)
-
-    const setMonthByPeriod = useStore(store, ({ actions: { setMonthByPeriod } }) => setMonthByPeriod)
-
-    return { setMonthByPeriod }
-  }
-
-  function DateStoreProvider({ children }: PropsWithChildren) {
-    const [store] = useState(() => buildStore())
-
-    return (
-      <DateStoreContext.Provider value={store}>
-        {children}
-      </DateStoreContext.Provider>
-    )
-  }
-
-  return {
-    Provider: DateStoreProvider,
-    useDate,
-    useDateActions,
-    useDateRange,
-    useDateRangeActions,
-    usePeriodAlignedActions,
-  }
-}
